Introduced simplified approach for get-file
The get-file command now supports --download-latest-into=/tmp (or --download-latest-into=/tmp/out2.txt) with auto-select of the first output provider when omitted, to obtain the latest file from the output folder. Useful for testing.

Also, just like put-file, the STORAGE_PROVIDER and REMOTE_PATH can be omitted and the values are fetched from the service definition.